Refactor files' names
